+++ trunk/shop/includes/lib/smarty/libs/plugins/shared.make_timestamp.php
@@ -22,8 +22,8 @@ function smarty_make_timestamp($string)
 
     } elseif (preg_match('/^\d{14}$/', $string)) {
         // it is mysql timestamp format of YYYYMMDDHHMMSS?
-        $time = mktime(substr($string, 8, 2),substr($string, 10, 2),substr($string, 12, 2),
-                       substr($string, 4, 2),substr($string, 6, 2),substr($string, 0, 4));
+        $time = mktime((int)substr($string, 8, 2),(int)substr($string, 10, 2),(int)substr($string, 12, 2),
+                       (int)substr($string, 4, 2),(int)substr($string, 6, 2),(int)substr($string, 0, 4));
 
     } elseif (is_numeric($string)) {
         // it is a numeric string, we handle it as timestamp
